Fix/types material

+++ src/main/java/co/edu/uceva/buildcheck/modules/materiales/DTO/MaterialDTO.java
@@ -13,7 +13,7 @@ public class MaterialDTO {
     String descripcion;
     String unidadMedida;
     Double precioUnitario;
-    Integer stockActual;
+    Double stockActual;
     LocalDateTime fechaCreacion;
     String usuarioCreador;
